+++ src/main/java/net/additionz/mixin/SweetBerryBushBlockMixin.java
@@ -1,0 +1,25 @@
+package net.additionz.mixin;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import net.additionz.AdditionMain;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.SweetBerryBushBlock;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+
+@Mixin(SweetBerryBushBlock.class)
+public class SweetBerryBushBlockMixin {
+
+    @Inject(method = "onEntityCollision", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;slowMovement(Lnet/minecraft/block/BlockState;Lnet/minecraft/util/math/Vec3d;)V"), cancellable = true)
+    private void onEntityCollisionMixin(BlockState state, World world, BlockPos pos, Entity entity, CallbackInfo info) {
+        if (AdditionMain.CONFIG.sneak_through_berries && ((LivingEntity) entity).isSneaking()) {
+            info.cancel();
+        }
+    }
+}