+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added emi fletching recipe category
 - Added teleport scroll
 - Added teleport potion
+- Added sneak through berries feature
 ### Fixed:
 - 
 ### Changed: